Refactoriza a paquete Planificadores
