+++ tareas/3/SantiagoDiego-UgaldeArmando/src/main/java/mx/unam/sistop/tarea3/ComparadorDePlanificadores.java
@@ -1,5 +1,7 @@
 package mx.unam.sistop.tarea3;
 
+import mx.unam.sistop.tarea3.planificadores.Planificadores;
+
 public class ComparadorDePlanificadores {
     private static final int RONDAS = 5;
