--- tareas/3/SantiagoDiego-UgaldeArmando/src/main/java/mx/unam/sistop/tarea3/Planificador.java
@@ -1,5 +1,0 @@
-package mx.unam.sistop.tarea3;
-
-public interface Planificador {
-    Resultado simular(CargaAleatoria cargaAleatoria);
-}
