+++ tareas/3/SantiagoDiego-UgaldeArmando/src/main/java/mx/unam/sistop/tarea3/planificadores/FirstComeFirstServed.java
@@ -1,4 +1,7 @@
-package mx.unam.sistop.tarea3;
+package mx.unam.sistop.tarea3.planificadores;
+
+import mx.unam.sistop.tarea3.CargaAleatoria;
+import mx.unam.sistop.tarea3.Resultado;
 
 public class FirstComeFirstServed implements Planificador {
     @Override
