+++ tareas/3/SantiagoDiego-UgaldeArmando/src/main/java/mx/unam/sistop/tarea3/planificadores/Planificador.java
@@ -1,0 +1,8 @@
+package mx.unam.sistop.tarea3.planificadores;
+
+import mx.unam.sistop.tarea3.CargaAleatoria;
+import mx.unam.sistop.tarea3.Resultado;
+
+public interface Planificador {
+    Resultado simular(CargaAleatoria cargaAleatoria);
+}
