+++ tareas/3/SantiagoDiego-UgaldeArmando/src/main/java/mx/unam/sistop/tarea3/planificadores/Planificadores.java
@@ -1,8 +1,11 @@
-package mx.unam.sistop.tarea3;
+package mx.unam.sistop.tarea3.planificadores;
+
+import mx.unam.sistop.tarea3.CargaAleatoria;
 
 public class Planificadores {
 
     public static void ejecutar(CargaAleatoria cargaAleatoria) {
         System.out.println(cargaAleatoria);
+
     }
 }
